agent/pyproject.toml
[project]
name = "wpilib-agent-tools"
version = "0.1.0"
description = "WPILib Agent Tools for FRC robot projects"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "robotpy-wpiutil",
    "robotpy-ntcore",
    "numpy",
    "matplotlib",
]

[project.scripts]
wpilib-agent-tools = "wpilib_agent_tools.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wpilib_agent_tools"]
